pyproject.toml
[tool.poetry]
name = "housing"
version = "0.1.0"
description = "Boston housing price analysis and prediction (ITMO EPML HW)"
authors = ["Artyom Kabanov"]
readme = "README.md"
packages = [{include = "housing", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
pandas = "2.2.3"
numpy = "1.26.4"
scikit-learn = "1.5.2"
matplotlib = "3.9.4"
seaborn = "0.13.2"
jupyter = "1.1.1"
ipykernel = "6.29.5"
dvc = "^3.66.1"
mlflow = "^3.10.0"
hydra-core = "^1.3.2"
clearml = "^1.16.5"

[tool.poetry.group.dev.dependencies]
ruff = "0.9.10"
mypy = "1.15.0"
bandit = "1.8.3"
pre-commit = "4.5.1"
pytest = "8.3.5"
pytest-cov = "6.0.0"
pandas-stubs = "^2.2.3"

# ── Ruff ──────────────────────────────────────────────────────────────────────
types-pyyaml = "^6.0.12.20250915"
[tool.ruff]
target-version = "py310"
line-length = 88
exclude = ["template/"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "S",    # flake8-bandit (security)
    "N",    # pep8-naming
]
ignore = [
    "S101",  # allow assert in tests
    "N803",  # allow uppercase X argument names (ML convention)
    "N806",  # allow uppercase X variable names (ML convention)
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101"]

[tool.ruff.lint.isort]
known-first-party = ["housing"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# ── MyPy ──────────────────────────────────────────────────────────────────────
[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
exclude = ["notebooks/", "template/"]

# ── Bandit ────────────────────────────────────────────────────────────────────
[tool.bandit]
exclude_dirs = ["tests", "notebooks", "template"]
skips = ["B101", "B301"]

# ── Pytest ────────────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=src/housing --cov-report=term-missing"

# ── Coverage ──────────────────────────────────────────────────────────────────
[tool.coverage.run]
source = ["src/housing"]
omit = ["tests/*"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
